italic-detector/requirements.txt
numpy>=1.20.0
scikit-learn>=1.0.0
pillow>=8.0.0
opencv-python>=4.5.0
pandas>=1.3.0
matplotlib>=3.4.0
PyMuPDF>=1.18.0
skl2onnx>=1.9.0
onnx>=1.10.0
onnxruntime>=1.8.0
click>=8.0.0
tqdm>=4.62.0